backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.4.2
python-dotenv==1.0.0
httpx==0.24.0
python-multipart==0.0.6
supabase==1.2.0
aiofiles==23.2.1
pytest==7.4.2
uvicorn
email-validator
bcrypt==4.0.1
pyjwt==2.8.0